Working on more memory management...
Might need to rethink this, but trying to have each struct own the
memory for its members, which is going to require more allocating and
copying of data to implement.

Currently, Object instances don't own the Expressions and Statements
they contain, leading to segfauls when the original Parser that created
them is deinitialized.